test: keep event replay coverage local to the arrakis hook
Revert all changes to the shared uniswap-v4 events test file and move
the dynamic fee pool Swap/ModifyLiquidity replay tests into
src/dex/uniswap-v4/hooks/arrakis/ to keep the PR scoped to the
integration.
